requirements.txt
couchbase==4.3.0
langchain-community==0.2.6
langchain-openai==0.1.14
langchain==0.2.6
tiktoken==0.7.0
python-dotenv==1.0.1
requests==2.32.3
